# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/control/{uid}/add": {
            "post": {
                "tags": [
                    "Control"
                ],
                "summary": "Adds a new control",
                "operationId": "9720ded00bf23eaa0cc8f26580e7e082",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Add control for uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 3
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "domains": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            "google.com",
                                            "yandex.ru",
                                            "site1.net"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "control response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Control_domain"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/control/{cid}": {
            "get": {
                "tags": [
                    "Control"
                ],
                "summary": "Get control by cid",
                "operationId": "b7ffcf595c9add4245ac0ec6f332ab51",
                "parameters": [
                    {
                        "name": "cid",
                        "in": "path",
                        "description": "Control to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 4
                    }
                ],
                "responses": {
                    "200": {
                        "description": "control response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Control_domain"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Control"
                ],
                "summary": "Updates a control",
                "operationId": "c00d506424568eca975c5d425d3729b7",
                "parameters": [
                    {
                        "name": "cid",
                        "in": "path",
                        "description": "cid to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "control": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "control": "google.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/control/{uid}/{cid}": {
            "get": {
                "tags": [
                    "Control"
                ],
                "summary": "Get control by uid",
                "operationId": "8ec1f5d3c2082682356edfd3b04f6610",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Control fetch by uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 3
                    },
                    {
                        "name": "cid",
                        "in": "path",
                        "description": "Control fetch by cid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 0
                    },
                    {
                        "name": "did",
                        "in": "query",
                        "description": "did",
                        "schema": {
                            "type": "string"
                        },
                        "example": 2
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "domain",
                        "schema": {
                            "type": "string"
                        },
                        "example": "site.com"
                    },
                    {
                        "name": "select",
                        "in": "query",
                        "description": "select",
                        "schema": {
                            "type": "string"
                        },
                        "example": "host_id,hidden,pos_settings"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "control response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Control_domain"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Control"
                ],
                "description": "deletes a single control ",
                "operationId": "4bc11c103dbe9fcb759d00030bcbcfc5",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 3
                    },
                    {
                        "name": "cid",
                        "in": "path",
                        "description": "cid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 0
                    }
                ],
                "responses": {
                    "204": {
                        "description": "control deleted"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/control/{uid}/{cid}/set_lr/{lr}": {
            "get": {
                "tags": [
                    "Control"
                ],
                "summary": "Set control lr by uid, cid",
                "operationId": "4e6458972547f9ddec10109400b6e1d1",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 4
                    },
                    {
                        "name": "cid",
                        "in": "path",
                        "description": "cid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 4
                    },
                    {
                        "name": "lr",
                        "in": "path",
                        "description": "lr",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "control response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Control_domain"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/control/{uid}/groups/": {
            "get": {
                "tags": [
                    "Control"
                ],
                "summary": "Get groups  list",
                "operationId": "cbcca4999a6199863a27d185dce705f7",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Groups fetch by uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 3
                    },
                    {
                        "name": "gid",
                        "in": "query",
                        "description": "gid",
                        "schema": {
                            "type": "string"
                        },
                        "example": "1,3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "control response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Control_domain"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Control_domain": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "did"
                        ],
                        "properties": {
                            "cid": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "uid": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "did": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "yid": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "metrika_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "yandex_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "host_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "lr": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "gid": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "pos_settings": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}